% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{add_presence}
\alias{add_presence}
\title{Add Presence Column}
\usage{
add_presence(df, cover_class_col = .data$`Cover Class`, absence_value = 0)
}
\arguments{
\item{df}{Tibble containing Cover Class Data}

\item{cover_class_col}{Column Name for Cover Class Data}

\item{absence_value}{Value for Absence -- Assumed to be Zero}
}
\value{
Provided tibble with column of logicals indicating presence / absence appended
}
\description{
Add Presence Column
}
\examples{
sagebrush \%>\%
  add_presence()
}
