% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{alluvial_plot}
\alias{alluvial_plot}
\title{Basic Alluvial Plots for Long Format Dataframes}
\usage{
alluvial_plot(.data, ..., alluvium_width = 1/12, stratum_width = 1/8)
}
\arguments{
\item{.data}{Dataframe from which the alluvial plot will be constructed}

\item{...}{Columns for the Alluvial Plot in Order; The first column will be used as the fill color}

\item{alluvium_width}{Width of the alluvium}

\item{stratum_width}{Width of the stratum}
}
\value{
ggplot2 object
}
\description{
Basic Alluvial Plots for Long Format Dataframes
}
\examples{
sagebrush \%>\%
  alluvial_plot(`Cover Class`, Fire)
}
