% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_functions.R
\name{braun_blanquet}
\alias{braun_blanquet}
\title{Braun-Blanquet Cutpoint Scheme}
\usage{
braun_blanquet()
}
\value{
Vector of Values Representing Braun-Blanquet Cover Class Cutpoint Scheme
}
\description{
Braun-Blanquet Cutpoint Scheme
}
\examples{
braun_blanquet()
}
