% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_functions.R
\name{daubenmire}
\alias{daubenmire}
\title{Daubenmire Cutpoint Scheme}
\usage{
daubenmire()
}
\value{
Vector of Values Representing Braun-Blanquet Cover Class Cutpoint Scheme
}
\description{
Daubenmire Cutpoint Scheme
}
\examples{
daubenmire()
}
