% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_functions.R
\name{domin}
\alias{domin}
\title{Domin Cutpoint Scheme}
\usage{
domin()
}
\value{
Vector of Values Representing Domin Cover Class Cutpoint Scheme
}
\description{
Domin Cutpoint Scheme
}
\examples{
domin()
}
