% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_functions.R
\name{hult_sernander}
\alias{hult_sernander}
\title{Hult-Sernander Cutpoint Scheme}
\usage{
hult_sernander()
}
\value{
Vector of Values Representing Hult-Sernander Cover Class Cutpoint Scheme
}
\description{
Hult-Sernander Cutpoint Scheme
}
\examples{
hult_sernander()
}
