% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{mosaic_plot}
\alias{mosaic_plot}
\title{Basic Mosaic Plots}
\usage{
mosaic_plot(df, factor1, factor2, fill)
}
\arguments{
\item{df}{Dataframe from which the mosaic plot will be constructed}

\item{factor1}{The first factor to be included in the mosaic plot}

\item{factor2}{The second factor to be included in the mosaic plot}

\item{fill}{The factor to be used for the fill color for the mosaic plot}
}
\value{

}
\description{
Basic Mosaic Plots
}
