% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{pivot_nps_data}
\alias{pivot_nps_data}
\title{NPS Data Format Transformation}
\usage{
pivot_nps_data(nps_data, covariate_cols = NULL)
}
\arguments{
\item{nps_data}{Tibble containing NPS Formatted Data}

\item{covariate_cols}{Columns which refer to covariates such as Year, Location, Grade, etc.}
}
\value{
Long format tibble
}
\description{
NPS Data Format Transformation
}
\examples{
\dontrun{
  pivot_nps_data(nps_data_ex, c(Year, Loc_Name, SCOSA))
}
}
