% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutpoint_functions.R
\name{validate_cutpoint}
\alias{validate_cutpoint}
\title{Validate a Cutpoint Vector / Scheme}
\usage{
validate_cutpoint(cutpoint_vector)
}
\arguments{
\item{cutpoint_vector}{A Vector of Decimal Cutpoints Bounded Between 0 and 1}
}
\value{

}
\description{
Ensures cutpoint values are:
\itemize{
\item Strictly Greater Than Zero
\item Strictly Less Than One
\item Distinct
\item Monotonically Increasing
}
}
\examples{
\dontrun{
validate_cutpoint(c(0.5)) # Okay - Remains Silent
validate_cutpoint(c(0.6, 0.4)) # Not Okay - Stops with an Error
}
}
