% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{ozab}
\alias{ozab}
\title{Fit OZAB Model}
\usage{
ozab(df, presence_formula, abundance_formula, cutpoint_scheme, ...)
}
\arguments{
\item{df}{Tibble containing Species, Cover Class, and Covariates}

\item{presence_formula}{Formula for Specifying Presence / Absence}

\item{abundance_formula}{Formula for Specifying Abundance}

\item{cutpoint_scheme}{Vector of Cutpoint Scheme Used}
}
\value{
Stan Object Fit
}
\description{
Fit OZAB Model
}
