% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_cover_class_by_covariate}
\alias{plot_cover_class_by_covariate}
\title{Plot Cumulative Empirical Proportions of Species by Covariate Bins}
\usage{
plot_cover_class_by_covariate(
  data,
  covariate,
  cover_class_col = .data$`Cover Class`
)
}
\arguments{
\item{data}{Dataframe in long format}

\item{covariate}{Name of Categorical Covariate to Plot By}

\item{cover_class_col}{Name of Cover Class column if it differs from Default 'Cover Class'}
}
\value{
ggplot2 object
}
\description{
First found in Irvine, Rodhouse, Keren 2016
}
\examples{
  plot_cover_class_by_covariate(sagebrush, Fire)
}
