% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence_absence_simulation_functions.R
\name{simulate_presence_absence_data}
\alias{simulate_presence_absence_data}
\title{Simulate Presence-Absence Data}
\usage{
simulate_presence_absence_data(
  S = 10,
  N = 100,
  spp_mean_presence = rnorm(S),
  spp_presence_correlation = rlkjcorr(1, S, eta = 1)
)
}
\arguments{
\item{S}{The Number of Species}

\item{N}{The Number of Plots or Sites Observed}

\item{spp_mean_presence}{The Gaussian latent species mean for prob. presence}

\item{presence_correlation}{An S by S correlation matrix between mean presence of species}

\item{detection_prob}{Detection probability can be either a single value representing detection probabilities across all species or a vector of length S containing species level detection probabilities}
}
\value{

}
\description{
Simulate Presence-Absence Data
}
\examples{
\dontrun{
simulate_presence_absence_data()
}
}
